Today's 3clear1fail baekjoons'

+++ -15829.c
@@ -1,5 +1,4 @@
 #include <stdio.h>
-#include <math.h>
 int main() {
     long long int sum=(long long int)0, mum=(long long int)1, i, j, n;
     char t;
@@ -8,7 +7,7 @@ int main() {
     for(i=0;i<n;i++) {
         scanf("%c", &t);
         t-=96;
-        mum=t;
+        mum=(long long int)t;
         for(j=0;j<i;j++) {
             mum*=(long long int)31;
             mum%=(long long int)1234567891;
